removed `adjugate` and `charpoly` from __init__, tests
